+++ src/Contracts/ArithmeticInterface.php
@@ -28,4 +28,28 @@ interface ArithmeticInterface
      * @return static
      */
     public function getBroadcastIp(int $cidr);
+
+    /**
+     * Get the Next Address
+     *
+     * @throws \Darsyn\IP\Exception\OverflowException
+     * @return static
+     */
+    public function next();
+
+    /**
+     * Get the Previous Address
+     *
+     * @throws \Darsyn\IP\Exception\OverflowException
+     * @return static
+     */
+    public function previous();
+
+    /**
+     * Get an Offset Address
+     *
+     * @throws \Darsyn\IP\Exception\OverflowException
+     * @return static
+     */
+    public function offset(int $offset);
 }